Don't use pipenv in tox anymore
Since the understreck package does not use any external dependencies (yet) the problem with the complex multi-python Pipfile.lock can be avoided by skipping pipenv for each tox test.

The dev deps is checked in one test running python 3.7. And the deployment to pypi is also happening in that job.